requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.1
google-generativeai
langchain-google-genai
chromadb
pypdf
langchain-community
python-docx
python-multipart 